Implement grid-based broad phase

+++ src/calculate_grid_size.rs
@@ -19,10 +19,10 @@ pub struct CalculateGridSize {
 impl CalculateGridSize {
     pub fn new(
         device: &Device,
-        grid_min_x: TypedBuffer<f32>,
-        grid_min_y: TypedBuffer<f32>,
-        grid_max_x: TypedBuffer<f32>,
-        grid_max_y: TypedBuffer<f32>,
+        grid_min_x: TypedBuffer<i32>,
+        grid_min_y: TypedBuffer<i32>,
+        grid_max_x: TypedBuffer<i32>,
+        grid_max_y: TypedBuffer<i32>,
         cell_size: TypedBuffer<f32>,
         grid_size: TypedBuffer<GridSize>,
     ) -> Self {
